Wrap each module code with a function wrapper
This is to keep top-level variables (defined with var) scoped to the
module rather than the global object and also provide access to the
global-looking 'module', 'exports', '__filename' and '__dirname'
variables. The 'main' module (i.e. the first to be loaded) is wrapped
using an Immediately Invoked Function Expression (IIFE), as it needs
to start executing immediately, whereas foreign modules should only be
executed on their corresponding require() call.