Add support for Node-style imports and exports (#1)
* Add ocamlformat and format everything

* Add flow AST traversal functions

These will be used to resolve 'require()' calls, which may be located
anywhere inside the JavaScript program.

* Transform 'require' calls by resolving the module paths

* Use normalised path strings

* Propagate the required module paths up

* Wrap each module code with a function wrapper

This is to keep top-level variables (defined with var) scoped to the
module rather than the global object and also provide access to the
global-looking 'module', 'exports', '__filename' and '__dirname'
variables. The 'main' module (i.e. the first to be loaded) is wrapped
using an Immediately Invoked Function Expression (IIFE), as it needs
to start executing immediately, whereas foreign modules should only be
executed on their corresponding require() call.

* Add preamble code at the beginning of the combined program

* Fix accidental reversal of statement order

* Fix wrong expression type for generated call params

* Switch to performing transformations at Syntax.exp level

This is to avoid passing the modified flow AST (which invariably had
invalid source code locations) to 'OfFLow'. Concrete and symbolic
execution works fine now.

* Change Loader_error to be like the others

* Do not preprocess module if there are no require() calls

This is a small optimisation to make the existing verification and
bi-abduction examples work as before.